Refactored again
