+++ packages/client/src/tests/util.tsx
@@ -69,6 +69,8 @@ export const validImageB64String =
   'iVBORw0KGgoAAAANSUhEUgAAAAgAAAACCAYAAABllJ3tAAAABHNCSVQICAgIfAhkiAAAABl0RVh0U29mdHdhcmUAZ25vbWUtc2NyZWVuc2hvdO8Dvz4AAAAXSURBVAiZY1RWVv7PgAcw4ZNkYGBgAABYyAFsic1CfAAAAABJRU5ErkJggg=='
 export const inValidImageB64String =
   'wee7dfaKGgoAAAANSUhEUgAAAAgAAAACCAYAAABllJ3tAAAABHNCSVQICAgIfAhkiAAAABl0RVh0U29mdHdhcmUAZ25vbWUtc2NyZWVuc2hvdO8Dvz4AAAAXSURBVAiZY1RWVv7PgAcw4ZNkYGBgAABYyAFsic1CfAAAAABJRU5ErkJggg=='
+export const natlsysadminToken =
+  'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzY29wZSI6WyJzeXNhZG1pbiIsIm5hdGxzeXNhZG1pbiIsImRlbW8iXSwiaWF0IjoxNjQ5MTUyMzE3LCJleHAiOjE2NDk3NTcxMTcsImF1ZCI6WyJvcGVuY3J2czphdXRoLXVzZXIiLCJvcGVuY3J2czp1c2VyLW1nbnQtdXNlciIsIm9wZW5jcnZzOmhlYXJ0aC11c2VyIiwib3BlbmNydnM6Z2F0ZXdheS11c2VyIiwib3BlbmNydnM6bm90aWZpY2F0aW9uLXVzZXIiLCJvcGVuY3J2czp3b3JrZmxvdy11c2VyIiwib3BlbmNydnM6c2VhcmNoLXVzZXIiLCJvcGVuY3J2czptZXRyaWNzLXVzZXIiLCJvcGVuY3J2czpjb3VudHJ5Y29uZmlnLXVzZXIiLCJvcGVuY3J2czp3ZWJob29rcy11c2VyIiwib3BlbmNydnM6Y29uZmlnLXVzZXIiXSwiaXNzIjoib3BlbmNydnM6YXV0aC1zZXJ2aWNlIiwic3ViIjoiNjIyZjgxYjQyY2Q1MzdiZjkxZGFhMTBiIn0.UEZnJp2s_9BdE2TIM1cdW2us7gWgjLqcrwXrHyu0_E6V1Mg5Vq8a_E8FB0Z1reXJjPQRAuNMjhs9iVaZYim4LS2LivhMk1XVmIfCDzYayT-PWT6nmoRDG2tJZoXUZbYFwTz0ef9IHrEgr1Xbpy_rJy_IFtb8uCu-j7TH7lGf6WzCIjSv6oq118IifUj9V4i9R6WoEZ7EsCcIQwAzf_3u8_qHCW7pFDfYf6nTTuQwLqX9ODpD2rBRKhfNH9TZFLqW9YzJZVQ6dZUBvyXpNRYppaHilNSi0geSl1uP5H7Sr98psV6iFsZiJ-EEwkKfMiprj37D3wPAV6o58FwZ_fwOCw'
 
 export const validateScopeToken = jwt.sign(
   { scope: ['validate'] },
@@ -1961,21 +1963,6 @@ export const userDetails = {
       }
     ]
   }
-}
-
-export const natlAdminUserDetails = {
-  userMgntUserID: '123',
-  language: 'en',
-  name: [
-    {
-      use: 'en',
-      firstNames: 'Taskin',
-      familyName: 'Ahmed'
-    },
-    { use: 'bn', firstNames: '', familyName: '' }
-  ],
-  role: 'NATIONAL_SYSTEM_ADMIN',
-  mobile: '01677701432'
 }
 
 export const mockUserResponseWithName = {
